refactor: rename codecs.py to person_codecs.py to avoid stdlib name conflict
