+++ .vscode/launch.json
@@ -26,7 +26,7 @@
             "name": "PyTest",
             "stopOnEntry": false,
             // "python": "${command:python.interpreterPath}",
-            "justMyCode": true,
+            "justMyCode": false,
             // "envFile": "${workspaceRoot}/.venv",
             "cwd": "${workspaceRoot}",
             "console": "internalConsole",
