+++ SESSION_RESUME.md
@@ -1,0 +1,135 @@
+# Session Resume — KWIC Prebuilt Refactor
+
+**Branch**: `speeches-download`  
+**Date**: 2026-04-10
+
+---
+
+## What Was Accomplished
+
+### 1. Committed Work (already in git)
+
+| Commit | Description |
+|--------|-------------|
+| `d38e0db` | `SpeechRepositoryFast` → `SpeechRepository` rename + startup `_align_with_dtm()` |
+| `208832f` / `7f7116b` | Added `docs/DESIGN.md` |
+| `3795cdc` | Initial `test_kwic_decode_regression.py` baseline |
+
+### 2. Uncommitted Changes (`git diff --stat HEAD` shows 14 files changed)
+
+These changes implement the **prebuilt `speech_index.feather`-based KWIC decode** — removing runtime codec lookups entirely:
+
+| File | Change |
+|------|--------|
+| `api_swedeb/workflows/prebuilt_speech_index/enrichment.py` | Added `wiki_id` to `SpeakerLookups` SQL query and `enrich_speech_rows` |
+| `api_swedeb/workflows/prebuilt_speech_index/build.py` | Added `wiki_id` to `_PYARROW_SCHEMA` and `_speech_rows_to_arrow` |
+| `api_swedeb/api/services/corpus_loader.py` | New `prebuilt_speech_index` lazy property (loads `speech_index.feather` indexed by `speech_id`) |
+| `api_swedeb/core/kwic/simple.py` | Rewrote `kwic_with_decode` to join prebuilt index; dropped `speech_index`+`codecs` params |
+| `api_swedeb/api/services/kwic_service.py` | Removed `codecs` parameter from `__init__` |
+| `api_swedeb/api/dependencies.py` | `get_kwic_service` no longer depends on `get_corpus_decoder` |
+| `tests/integration/test_kwic_decode_regression.py` | Updated fixture to use real data path; updated call signatures |
+| `tests/test_data/v1.4.1/speeches/bootstrap_corpus/speech_index.feather` | **NEW FILE** — minimal fixture generated from DTM document_index + codecs |
+
+---
+
+## Current Test State
+
+### Passing ✅
+- `tests/integration/test_kwic_decode_regression.py` — **14/14 pass**
+
+### Failing ❌ — Two categories:
+
+#### A. Tests in `test_kwic_core.py` using OLD `kwic_with_decode` signature
+
+These tests call `kwic_with_decode(..., speech_index=..., codecs=...)` — the old API.
+They need to be updated to use `prebuilt_speech_index=loader.prebuilt_speech_index`.
+
+Affected tests:
+- `test_simple_kwic_with_decode_results_for_various_setups` (6 parametrized cases)
+- `test_kwic_with_decode`
+- `test_kwic_with_decode_returns_enriched_data`
+- `test_kwic_with_decode_multiprocess`
+
+**Also** `test_bug_kwic_fails_when_lemmatized_is_true` in `test_kwic.py` calls `KWICService(loader, person_codecs)` — now it's `KWICService(loader)` only.
+
+**Fix**: Update these tests to use the new signature. The `_kwic_loader` fixture pattern from `test_kwic_decode_regression.py` can be reused, or pass `corpus_loader.prebuilt_speech_index`.
+
+#### B. Tests with wrong expected result counts
+
+- `test_kwic_filter_by_chamber[sverige-ek-225]` — expects 225, gets different count
+- `test_kwic_filter_by_chamber[sverige-chambers3-267]` — expects 267, gets different count
+- `test_kwic_speech_id_in_search_results` — gets 0 results for `kärnkraft` (may be a test corpus limitation)
+
+These may be pre-existing data issues (test CWB corpus vs expectation mismatch) unrelated to the current refactor. **Check git blame** to see if they were passing before.
+
+---
+
+## Side Note: document_name Mismatch Warning
+
+When `corpus_loader` loads the test data, `_align_with_dtm()` logs:
+
+```
+WARNING: 443 speech_ids have mismatched document_name between DTM and prebuilt;
+  first 5: [('i-Dq...', 'prot-1970--ak--029_001', 'prot-1970--ak--029_1'), ...]
+```
+
+This is because the **minimal `speech_index.feather`** we generated for tests derived `document_name` from the DTM with a different naming convention (`_001` vs `_1`). The warn is benign for now but worth fixing in the test fixture generation script.
+
+---
+
+## Next Steps (Priority Order)
+
+### 1. Fix failing `test_kwic_core.py` tests (signature update)
+
+The tests need to get a `prebuilt_speech_index` instead of `speech_index + codecs`.
+Options:
+- Add a module-scoped `prebuilt_speech_index` fixture to `conftest.py`:
+  ```python
+  @pytest.fixture(scope="module")
+  def prebuilt_speech_index(corpus_loader: CorpusLoader) -> pd.DataFrame:
+      return corpus_loader.prebuilt_speech_index
+  ```
+- Then update all calls from:
+  ```python
+  simple.kwic_with_decode(corpus, opts, speech_index=speech_index, codecs=person_codecs, ...)
+  ```
+  to:
+  ```python
+  simple.kwic_with_decode(corpus, opts, prebuilt_speech_index=prebuilt_speech_index, ...)
+  ```
+
+### 2. Fix `test_bug_kwic_fails_when_lemmatized_is_true` 
+
+Update `KWICService(loader, person_codecs)` → `KWICService(loader)`.
+
+### 3. Investigate chamber/speech_id count failures
+
+Check if `test_kwic_filter_by_chamber` and `test_kwic_speech_id_in_search_results` were
+already failing before this refactor by checking `git stash` + rerun, or checking git blame.
+
+### 4. Commit everything
+
+```bash
+git add -A
+git commit -m "refactor: kwic_with_decode uses prebuilt speech_index instead of DTM + codecs"
+```
+
+### 5. Rebuild corpus (when ready) to get `wiki_id` in real feather
+
+The existing `data/v1.4.1/speeches/bootstrap_corpus/speech_index.feather` was built before
+the `wiki_id` enrichment change. Run the build pipeline to regenerate with `wiki_id`.
+
+---
+
+## Key File Locations
+
+```
+api_swedeb/core/kwic/simple.py            ← kwic_with_decode (new prebuilt API)
+api_swedeb/api/services/corpus_loader.py  ← prebuilt_speech_index property (line ~88)
+api_swedeb/api/services/kwic_service.py   ← KWICService (no codecs param)
+api_swedeb/api/dependencies.py            ← get_kwic_service (no get_corpus_decoder)
+api_swedeb/workflows/prebuilt_speech_index/enrichment.py  ← wiki_id added
+api_swedeb/workflows/prebuilt_speech_index/build.py       ← wiki_id in schema
+tests/integration/test_kwic_decode_regression.py  ← 14 passing regression tests
+tests/test_data/v1.4.1/speeches/bootstrap_corpus/speech_index.feather  ← NEW test fixture
+```
