+++ api_swedeb/api/services/corpus_loader.py
@@ -15,7 +15,7 @@ from typing import Optional
 
 import pandas as pd
 
-from api_swedeb.core import codecs as md
+from api_swedeb.core import person_codecs as md
 from api_swedeb.core.configuration import ConfigValue
 from api_swedeb.core.load import load_dtm_corpus, load_speech_index
 from api_swedeb.core.speech_repository import SpeechRepository
