+++ tests/api_swedeb/core/test_codecs.py
@@ -11,7 +11,7 @@ from unittest.mock import MagicMock, patch
 import pandas as pd
 import pytest
 
-from api_swedeb.core.codecs import Codec, Codecs, MultiplePartyAbbrevsHook, PersonCodecs
+from api_swedeb.core.person_codecs import Codec, Codecs, MultiplePartyAbbrevsHook, PersonCodecs
 
 # pylint: disable=protected-access, too-many-public-methods
 
@@ -547,7 +547,7 @@ class TestBaseCodecs:
 class TestCodecs:
     """Test cases for Codecs class."""
 
-    @patch('api_swedeb.core.codecs.ConfigValue')
+    @patch('api_swedeb.core.person_codecs.ConfigValue')
     def test_codecs_initialization_default(self, mock_config_value):
         """Test PersonCodecs initialization with default configuration."""
         mock_specification = {"tables": {}, "codecs": []}
@@ -570,7 +570,7 @@ class TestCodecs:
 class TestPersonCodecs:
     """Test cases for PersonCodecs class."""
 
-    @patch('api_swedeb.core.codecs.ConfigValue')
+    @patch('api_swedeb.core.person_codecs.ConfigValue')
     def test_person_codecs_initialization(self, mock_config_value):
         """Test PersonCodecs initialization merges configurations."""
         mock_mappings = {
@@ -675,7 +675,7 @@ class TestPersonCodecs:
         expected: str = "https://www.wikidata.org/wiki/Q123456"
         assert str(result) == expected
 
-    @patch('api_swedeb.core.codecs.ConfigValue')
+    @patch('api_swedeb.core.person_codecs.ConfigValue')
     def test_person_wiki_link_unknown_value(self, mock_config_value):
         """Test person_wiki_link with unknown value."""
         mock_config_value.return_value.resolve.return_value = "Unknown Speaker"
@@ -687,7 +687,7 @@ class TestPersonCodecs:
         """Test person_wiki_link with pandas Series."""
         wiki_ids = pd.Series(["Q123", "Q456", "unknown"])
 
-        with patch('api_swedeb.core.codecs.ConfigValue') as mock_config_value:
+        with patch('api_swedeb.core.person_codecs.ConfigValue') as mock_config_value:
             mock_config_value.return_value.resolve.return_value = "Unknown Speaker"
 
             result = PersonCodecs.person_wiki_link(wiki_ids)
@@ -699,7 +699,7 @@ class TestPersonCodecs:
 
             pd.testing.assert_series_equal(result, expected)
 
-    @patch('api_swedeb.core.codecs.ConfigValue')
+    @patch('api_swedeb.core.person_codecs.ConfigValue')
     def test_speech_link_single_document(self, mock_config_value):
         """Test speech_link with single document."""
         mock_config_value.return_value.resolve.return_value = "https://example.com/"
@@ -709,7 +709,7 @@ class TestPersonCodecs:
         assert isinstance(result, str)
         assert result == expected
 
-    @patch('api_swedeb.core.codecs.ConfigValue')
+    @patch('api_swedeb.core.person_codecs.ConfigValue')
     def test_speech_link_series(self, mock_config_value):
         """Test speech_link with pandas Series."""
         mock_config_value.return_value.resolve.return_value = "https://example.com/"
