+++ tests/integration/test_cwb_integration.py
@@ -4,7 +4,7 @@ import ccc
 import pandas as pd
 import pytest
 
-from api_swedeb.core.codecs import PersonCodecs
+from api_swedeb.core.person_codecs import PersonCodecs
 from api_swedeb.core.cwb import compiler
 from api_swedeb.core.cwb.utility import CorpusAttribs
